+++ HISTORY.md
@@ -25,6 +25,7 @@
 * Add support for `BaseCache` keyword arguments passed along with a backend instance
 * Fix issue with cache headers not being used correctly if `cache_control=True` is used with an `expire_after` value
 * Fix license metadata as shown on PyPI
+* Fix `CachedResponse` serialization behavior when using stdlib `pickle` in a custom serializer
 
 ## 0.8.1 (2021-09-15)
 * Redact `ingored_parameters` from `CachedResponse.url` (if used for credentials or other sensitive info)